+++ bundle.sh
@@ -1,0 +1,5 @@
+#! /bin/bash
+cp TraXile.Updater/bin/Release/TraXile.Updater.exe TraXile/bin/Release/
+rm -r TraXile/bin/Release/app.publish
+rm -r TraXile/bin/Release/TraXile.application
+