+++ clear_bin.sh
@@ -1,0 +1,5 @@
+#! /bin/bash
+rm -r TraXile/bin
+rm -r TraXile.Updater/bin
+rm -r Setup/Debug
+rm -r Setup/Release